data-scaling/requirements.txt
numpy==1.23.5
Pillow==10.0.0
pydicom==2.3.1
scipy==1.10.1